FlaskCacheApp/requirements.txt
blinker==1.8.2
cachelib==0.9.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
Faker==30.8.0
Flask==3.0.3
Flask-Caching==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.1.1
six==1.16.0
typing_extensions==4.12.2
Werkzeug==3.0.4
